fix: 상태관리 로직 변경
